Refactor Forum, Category and Doc
