+++ app/controllers/admin/base_controller.rb
@@ -1,0 +1,7 @@
+class Admin::BaseController < ApplicationController
+
+  layout 'admin'
+  before_action :authenticate_user!
+  before_action :verify_admin
+
+end
