+++ app/controllers/forums_controller.rb
@@ -17,14 +17,7 @@
 #
 
 class ForumsController < ApplicationController
-  #cache_sweeper :forum_sweeper, :only => [:update, :create, :destroy]
 
-  before_action :authenticate_user!, :only => ['new', 'edit', 'update', 'create', 'destroy']
-  before_action :verify_admin, :only => ['new', 'edit', 'update', 'create', 'destroy']
-  layout 'admin', :only => ['new', 'edit', 'update', 'create']
-
-  # GET /forums
-  # GET /forums.xml
   def index
     @page_title = t(:community, default: "Community")
 
@@ -40,78 +33,8 @@ class ForumsController < ApplicationController
     end
   end
 
-  # GET /forums/1
   def show
     redirect_to topics_path(:forum_id => params[:id])
   end
 
-  # GET /forums/new
-  def new
-    @forum = Forum.new
-  end
-
-  # GET /forums/1;edit
-  def edit
-    @forum = Forum.find(params[:id])
-  end
-
-  # POST /forums
-  # POST /forums.xml
-  def create
-    @forum = Forum.new(forum_params)
-
-    respond_to do |format|
-      if @forum.save
-        flash[:notice] = 'Forum was successfully created.'
-        format.html { redirect_to admin_communities_path }
-        format.xml  { head :created, :location => forum_url(@forum) }
-      else
-        format.html { render :new }
-        format.xml  { render :xml => @forum.errors.to_xml }
-      end
-    end
-  end
-
-  # PUT /forums/1
-  # PUT /forums/1.xml
-  def update
-    @forum = Forum.find(params[:id])
-
-    respond_to do |format|
-      if @forum.update(forum_params)
-        flash[:notice] = 'Forum was successfully updated.'
-        format.html { redirect_to admin_communities_path }
-        format.xml  { head :ok }
-      else
-        format.html { render :edit }
-        format.xml  { render :xml => @forum.errors.to_xml }
-      end
-    end
-  end
-
-  # DELETE /forums/1
-  # DELETE /forums/1.xml
-  def destroy
-    @forum = Forum.find(params[:id])
-    @forum.destroy
-
-    respond_to do |format|
-      format.html { redirect_to admin_communities_path }
-      format.js
-      format.xml  { head :ok }
-    end
-  end
-
-  private
-
-  def forum_params
-    params.require(:forum).permit(
-      :name,
-      :description,
-      :layout,
-      :private,
-      :allow_topic_voting,
-      :allow_post_voting
-    )
-  end
 end
