+++ app/views/admin/_post.html.erb
@@ -9,7 +9,7 @@
 			<%= post_message(post,true) %>
 			<ul class="dropdown-menu dropdown-menu-left" role="menu">
 				<li><%= link_to 'Edit', edit_post_path(post), remote: true %></li>
-				<li><%= link_to 'Turn post into content', new_doc_path(post_id: post.id, lang: I18n.locale) %></li>
+				<li><%= link_to 'Turn post into content', new_admin_doc_path(post_id: post.id, lang: I18n.locale) %></li>
 			</ul>
 		</span>
