+++ app/views/admin/_ticket.html.erb
@@ -1,7 +1,7 @@
 <div id="ticket-info" class="no-side-pad row">
 
   <div class="col-md-4 hidden-xs hidden-sm" id="left-col-user-info">
-    <%= render 'admin/user_info' %>
+    <%= render partial: 'admin/users/user_info' %>
   </div>
 
   <div id="ticket-controls" class="col-md-8 col-xs-12 col-lg-8">
