+++ app/views/admin/categories/_form.html.erb
