+++ app/views/admin/categories/destroy.js.erb
