+++ app/views/admin/categories/edit.html.erb
@@ -7,7 +7,7 @@
 
 <%= locale_select if AppSettings['i18n.available_locales'].count > 1 %>
 
-<%= bootstrap_form_for @category, url: category_path(id: @category.id), label_errors: true, method: 'patch' do |f| %>
+<%= bootstrap_form_for @category, url: admin_category_path(id: @category.id), label_errors: true, method: 'patch' do |f| %>
   <% Globalize.with_locale(params['lang']) do %>
     <%= f.alert_message "Please fix the errors below." %>
     <%= render :partial => 'form', :object => @category, :locals => {:f => f} %>
