+++ app/views/admin/categories/index.html.erb
@@ -13,8 +13,8 @@ $(document).ready(function(){
 <div class="admin-header">
   <h2><%= t(:content, default: 'Content') %>
     <span class="admin-options text-right pull-right">
-    <%= link_to t(:new_category, default: "New Category"), new_category_path(lang: I18n.locale), class: 'btn btn-primary' %>
-    <%= link_to t(:new_content, default: "New Content"), new_doc_path(lang: I18n.locale), class: 'btn btn-primary' %>
+    <%= link_to t(:new_category, default: "New Category"), new_admin_category_path(lang: I18n.locale), class: 'btn btn-primary' %>
+    <%= link_to t(:new_content, default: "New Content"), new_admin_doc_path(lang: I18n.locale), class: 'btn btn-primary' %>
     </span>
   </h2>
 </div>
