+++ app/views/admin/categories/show.html.erb
@@ -1,8 +1,8 @@
 <h2><%= "#{@category.name}" %></h2>
 
 <div class="admin-options text-right">
-  <%= link_to t(:new_category, default: "New Category"), new_category_path(lang: I18n.locale), class: 'btn btn-primary' %>
-  <%= link_to t(:new_content, default: "New Content"), new_doc_path(category_id: @category.id, lang: I18n.locale), class: 'btn btn-primary' %>
+  <%= link_to t(:new_category, default: "New Category"), new_admin_category_path(lang: I18n.locale), class: 'btn btn-primary' %>
+  <%= link_to t(:new_content, default: "New Content"), new_admin_category_doc_path(category_id: @category.id, lang: I18n.locale), class: 'btn btn-primary' %>
 </div>
 <div class="row">
   <div class="col-md-12">
@@ -12,7 +12,7 @@
       <% @docs.each do |doc| %>
       <tr id="doc-<%= doc.id %>" data-obj-id="<%= "#{doc.id}" %>" data-obj='doc' class="item">
         <td>
-          <span class="more-important"><%= link_to doc.title, edit_category_doc_path(doc.category.id, doc.id, lang: I18n.locale)  %> <%= " - #{I18n.t('featured_content', default: 'Featured')}" if doc.front_page == true %> <%= " - #{I18n.t('draft_content', default: 'Draft')}" if doc.active == false %></span>
+          <span class="more-important"><%= link_to doc.title, edit_admin_category_doc_path(doc.category.id, doc.id, lang: I18n.locale)  %> <%= " - #{I18n.t('featured_content', default: 'Featured')}" if doc.front_page == true %> <%= " - #{I18n.t('draft_content', default: 'Draft')}" if doc.active == false %></span>
           <%= i18n_icons(doc) if AppSettings['i18n.available_locales'].count > 1 %>
           <br/>
           <span class="less-important"><%= I18n.t('written_on', date_written: doc.created_at.strftime("%m/%d/%Y"), default: 'Written on') %> <%= doc.updated_at == doc.created_at ? "" : " | #{I18n.t("content_last_updated", date_updated: doc.updated_at.strftime("%m/%d/%Y"))}" %></span>
@@ -21,8 +21,8 @@
           <div class="btn-group">
             <span id="row-<%= doc.id %>" data-toggle="dropdown" aria-expanded="false" class='btn dropdown-toggle glyphicon glyphicon-align-justify'></span>
             <ul class="dropdown-menu dropdown-menu-right" role="menu">
-              <li><%= link_to t(:edit, default: 'Edit'), edit_category_doc_path(doc.category.id, doc.id, lang: I18n.locale) %></li>
-              <li><%= link_to t(:delete, default: 'Delete'), doc, data: {confirm: t(:delete_confirm, default: 'Please confirm you really want to DELETE this')}, method: :delete, remote: true, class: 'less-important' %></li>
+              <li><%= link_to t(:edit, default: 'Edit'), edit_admin_category_doc_path(doc.category.id, doc.id, lang: I18n.locale) %></li>
+              <li><%= link_to t(:delete, default: 'Delete'), admin_doc_path(doc.id), data: {confirm: t(:delete_confirm, default: 'Please confirm you really want to DELETE this')}, method: :delete, remote: true, class: 'less-important' %></li>
               <li><%= link_to t(:view_on_site, default: 'View on Site'), category_doc_path(doc.category, doc), {target: "_blank"} %></li>
             </ul>
           </div>
