+++ app/views/admin/forums/_forum.html.erb
@@ -13,8 +13,8 @@
     <div class="btn btn-group">
       <span id="row-<%= forum.id %>" data-toggle="dropdown" aria-expanded="false" class='dropdown-toggle glyphicon glyphicon-align-justify'></span>
       <ul class="dropdown-menu dropdown-menu-right" role="menu">
-        <li><%= link_to t(:edit, default:'Edit'), edit_forum_path(forum) %></li>
-        <li><%= link_to t(:delete, default:'Delete'), forum, data: {confirm: 'Please confirm you really want to DELETE this forum and all discussions'}, method: :delete, remote: true, class: 'less-important' %></li>
+        <li><%= link_to t(:edit, default:'Edit'), edit_admin_forum_path(forum) %></li>
+        <li><%= link_to t(:delete, default:'Delete'), admin_forum_path(forum), data: {confirm: 'Please confirm you really want to DELETE this forum and all discussions'}, method: :delete, remote: true, class: 'less-important' %></li>
       </ul>
     </div>
   </td>
