+++ app/views/admin/forums/destroy.js
