+++ app/views/admin/forums/index.html.erb
@@ -1,5 +1,5 @@
 <h2>Admin Communities <span class="admin-options pull-right">
-  <%= link_to t(:new_community, default: "Create New Community"), new_forum_path, class: 'btn btn-primary' %>
+  <%= link_to t(:new_community, default: "Create New Community"), new_admin_forum_path, class: 'btn btn-primary' %>
 </span></h2>
 
 <div class="row">
