--- app/views/forums/edit.html.erb
@@ -1,6 +1,0 @@
-<h2><%= t(:edit, default: 'Edit') %>: <%= @forum.name %></h2>
-
-<%= bootstrap_form_for(@forum) do |f| %>
-  <%= render :partial => 'form', :locals => {:f => f} -%>
-  <%= link_to t(:delete, default: 'Delete'), @forum, data: {confirm: 'Please confirm you really want to DELETE this forum and all discussions'}, method: :delete, class: 'btn btn-danger  less-important' %>
-<% end %>
