+++ app/views/layouts/_admin_header.html.erb
@@ -20,8 +20,8 @@
           <%= menu_group pull: :right do %>
               <%#= menu_item "Dashboard", admin_dashboard_path %>
               <%= menu_item t(:discussions, default: "Discussions"), admin_tickets_path %>
-              <%= menu_item t(:content, default: "Content"), admin_knowledgebase_path, class:'kblink' if Settings.knowledgebase %>
-              <%= menu_item t(:communities, default: "Communities"), admin_communities_path if Settings.forums %>
+              <%= menu_item t(:content, default: "Content"), admin_categories_path, class:'kblink' if Settings.knowledgebase %>
+              <%= menu_item t(:communities, default: "Communities"), admin_forums_path if Settings.forums %>
               <%= menu_item t(:app_store, default: "App Store"), "http://helpy.io/store/" %>
           <% end %>
